+++ include/Planet.h
@@ -10,10 +10,10 @@ class Planet : public AstroObject {
 	private:
 		sf::CircleShape circle;
 	public:
-		Planet(long long x, long long y, float _radius, sf::Color _colour);
+		Planet(double x, double y, float _radius, sf::Color _colour);
 		Planet();
 		void update();
-		//void render(sf::RenderWindow &window, sf::Vector2<long long> view, sf::VideoMode screen);
+		//void render(sf::RenderWindow &window, sf::Vector2<double> view, sf::VideoMode screen);
 };
 
 #endif
