+++ src/SolarSystem.cpp
@@ -1,12 +1,12 @@
 #include <SolarSystem.h>
 
-SolarSystem::SolarSystem(long long _radius) {
+SolarSystem::SolarSystem(double _radius) {
 	radius = _radius;
 }
 
 SolarSystem::SolarSystem() : SolarSystem(1000000) {
 }
 
-long long SolarSystem::getRadius() {
+double SolarSystem::getRadius() {
 	return radius;
 }
