+++ src/Sun.cpp
@@ -1,6 +1,6 @@
 #include <Sun.h>
 
-Sun::Sun(long long x, long long y, float _radius, sf::Color _colour) : AstroObject(x, y, _radius, _colour) {
+Sun::Sun(double x, double y, float _radius, sf::Color _colour) : AstroObject(x, y, _radius, _colour) {
 	//circle.setRadius(_radius);
 	//circle.setPointCount(100);
 	//circle.setFillColor(_colour);
@@ -14,7 +14,7 @@ void Sun::update() {
 	//circle.setPosition(x - getRadius(), y - getRadius());
 }
 
-//void Sun::render(sf::RenderWindow &window, sf::Vector2<long long> view, sf::VideoMode screen) {
+//void Sun::render(sf::RenderWindow &window, sf::Vector2<double> view, sf::VideoMode screen) {
 	//circle.setPosition((screen.width / 2) + (getX() - view.x) - getRadius(), (screen.height / 2) + (getY() - view.y) - getRadius());
 	//window.draw(circle);
 //}
